[alpha_factory] factor shared config helpers
