+++ src/utils/config.py
@@ -5,32 +5,17 @@ from __future__ import annotations
 
 import logging
 import os
-from pathlib import Path
 from typing import Any, Optional, cast
 
-from alpha_factory_v1.utils.env import _load_env_file
+from alpha_factory_v1.utils.config_common import (
+    SettingsBase,
+    _load_dotenv,
+    _prefetch_vault,
+)
 
 from pydantic import Field
-from pydantic_settings import BaseSettings, SettingsConfigDict
 
 _log = logging.getLogger(__name__)
-
-
-def _load_dotenv(path: str = ".env") -> None:
-    """Load default variables from ``path`` when available.
-
-    Key/value pairs from the file are added to :mod:`os.environ` only if the
-    variable is not already defined.
-
-    Args:
-        path: Location of the ``.env`` file.
-    """
-
-    if Path(path).is_file():
-        for k, v in _load_env_file(path).items():
-            os.environ.setdefault(k, v)
-
-
 
 
 def get_secret(name: str, default: Optional[str] = None) -> Optional[str]:
@@ -96,26 +81,6 @@ def get_secret(name: str, default: Optional[str] = None) -> Optional[str]:
     return os.getenv(name, default)
 
 
-def _prefetch_vault() -> None:
-    """Populate environment secrets from HashiCorp Vault if configured."""
-    if "VAULT_ADDR" in os.environ:
-        try:  # pragma: no cover - optional dependency
-            import importlib
-
-            hvac = importlib.import_module("hvac")
-
-            addr = os.environ["VAULT_ADDR"]
-            token = os.getenv("VAULT_TOKEN")
-            secret_path = os.getenv("OPENAI_API_KEY_PATH", "OPENAI_API_KEY")
-            client = hvac.Client(url=addr, token=token)
-            data = client.secrets.kv.read_secret_version(path=secret_path)
-            value = data["data"]["data"].get("OPENAI_API_KEY")
-            if value:
-                os.environ["OPENAI_API_KEY"] = value
-        except Exception as exc:  # noqa: BLE001
-            _log.warning("Vault lookup failed: %s", exc)
-
-
 def init_config(env_file: str = ".env") -> None:
     """Load environment variables and refresh :data:`CFG`."""
 
@@ -125,22 +90,13 @@ def init_config(env_file: str = ".env") -> None:
     CFG = Settings()
 
 
-
-
-class Settings(BaseSettings):
+class Settings(SettingsBase):
     """Environment-driven configuration."""
 
     openai_api_key: Optional[str] = Field(default=None, alias="OPENAI_API_KEY")
     offline: bool = Field(default=False, alias="AGI_INSIGHT_OFFLINE")
     bus_port: int = Field(default=6006, alias="AGI_INSIGHT_BUS_PORT")
     ledger_path: str = Field(default="./ledger/audit.db", alias="AGI_INSIGHT_LEDGER_PATH")
-
-    model_config: SettingsConfigDict = {
-        "env_file": ".env",
-        "extra": "ignore",
-        "populate_by_name": True,
-        "env_prefix": "",
-    }
 
     def __init__(self, **data: Any) -> None:  # pragma: no cover - exercised in tests
         super().__init__(**data)
@@ -149,13 +105,6 @@ class Settings(BaseSettings):
         if not self.openai_api_key:
             _log.warning("OPENAI_API_KEY missing – offline mode enabled")
             self.offline = True
-
-    def __repr__(self) -> str:  # pragma: no cover - trivial
-        data = self.model_dump()
-        for k in tuple(data):
-            if any(s in k.lower() for s in ("token", "key", "password")) and data[k]:
-                data[k] = "***"
-        return f"Settings({data})"
 
 
 CFG = Settings()
